requirements.txt
pyyaml
ollama
python-docx
pymupdf
openpyxl
python-pptx
flask
chardet
httpx
